feat: rrs count
